Finish post module

+++ crud/src/routes/posts/posts.controller.ts
@@ -3,11 +3,8 @@ import { PostsService } from './posts.service'
 import { Auth } from 'src/shared/decorators/auth.decorator'
 import { AuthType, ConditionalGuard } from 'src/shared/constants/auth.constant'
 import { AuthenticationGuard } from 'src/shared/guards/authentication.guard'
-export interface Post {
-  id: string
-  title: string
-  content: string
-}
+import { ActiveUser } from 'src/shared/decorators/active-user.decorator'
+import { CreatePostBodyDTO, DeletePostBodyDTO, GetPostItemDTO, UpdatePostBodyDTO } from './post.dto'
 
 @Controller('posts')
 export class PostsController {
@@ -15,27 +12,33 @@ export class PostsController {
   @Get()
   @Auth([AuthType.Bearer, AuthType.APIKey], { conditional: ConditionalGuard.And })
   // @UseGuards(AuthenticationGuard)
-  getPosts() {
-    return this.postsService.getPosts()
+  async getPosts(@ActiveUser('userId') userId: number) {
+    return this.postsService.getPosts(userId).then((posts) => {
+      return posts.map((post) => new GetPostItemDTO(post))
+    })
   }
 
   @Post()
-  createPost(@Body() body: Post) {
-    return this.postsService.createPost(body)
+  @Auth([AuthType.Bearer])
+  async createPost(@Body() body: CreatePostBodyDTO, @ActiveUser('userId') userId: number) {
+    return new CreatePostBodyDTO(await this.postsService.createPost(userId, body))
   }
 
   @Get(':id')
-  getPost(@Param('id') id: string) {
-    return this.postsService.getPost(id)
+  @Auth([AuthType.Bearer])
+  async getPost(@Param('id') id: string) {
+    return new CreatePostBodyDTO(await this.postsService.getPost(Number(id)))
   }
 
   @Put(':id')
-  updatePost(@Param('id') id: string, @Body() body: any) {
-    return this.postsService.updatePost(id, body)
+  @Auth([AuthType.Bearer])
+  async updatePost(@Param('id') id: string, @Body() body: UpdatePostBodyDTO, @ActiveUser('userId') userId: number) {
+    return new UpdatePostBodyDTO(await this.postsService.updatePost({ postId: Number(id), userId, body }))
   }
 
   @Delete(':id')
-  deletePost(@Param('id') id: string) {
-    return this.postsService.deletePost(id)
+  @Auth([AuthType.Bearer])
+  async deletePost(@Param('id') id: string, @ActiveUser('userId') userId: number) {
+    return new DeletePostBodyDTO(await this.postsService.deletePost({ postId: Number(id), userId }))
   }
 }
